refactor: extract KatexMath into shared module
Moves the KaTeX rendering component from markdownRenderer.tsx into its
own file so LatexOutput can compose it rather than duplicating the
load/render/sanitize logic.
