+++ src/vs/workbench/contrib/positronNotebook/browser/notebookCells/LatexOutput.tsx
@@ -10,11 +10,9 @@ import './LatexOutput.css';
 import React from 'react';
 
 // Other dependencies.
-import { importAMDNodeModule } from '../../../../../amdX.js';
 import { getWindow } from '../../../../../base/browser/dom.js';
-import { safeSetInnerHtml } from '../../../../../base/browser/domSanitize.js';
-import { allowedMarkdownHtmlTags, allowedMarkdownHtmlAttributes } from '../../../../../base/browser/markdownRenderer.js';
 import { MarkedKatexSupport } from '../../../markdown/browser/markedKatexSupport.js';
+import { KatexMath } from './KatexMath.js';
 import { normalizeLatex } from './normalizeLatex.js';
 
 /**
@@ -27,45 +25,16 @@ export function LatexOutput({ content }: { content: string }) {
 	const containerRef = React.useRef<HTMLDivElement>(null);
 
 	React.useEffect(() => {
-		const container = containerRef.current;
-		if (!container) {
-			return;
+		if (containerRef.current) {
+			MarkedKatexSupport.ensureKatexStyles(getWindow(containerRef.current));
 		}
+	}, []);
 
-		let cancelled = false;
+	const latex = React.useMemo(() => normalizeLatex(content), [content]);
 
-		const latex = normalizeLatex(content);
-
-		MarkedKatexSupport.ensureKatexStyles(getWindow(container));
-
-		importAMDNodeModule<typeof import('katex').default>('katex', 'dist/katex.min.js').then(katex => {
-			if (cancelled) {
-				return;
-			}
-
-			try {
-				const html = katex.renderToString(latex, {
-					throwOnError: false,
-					displayMode: true,
-				});
-				const sanitizerConfig = MarkedKatexSupport.getSanitizerOptions({
-					allowedTags: allowedMarkdownHtmlTags,
-					allowedAttributes: allowedMarkdownHtmlAttributes,
-				});
-				safeSetInnerHtml(container, html, sanitizerConfig);
-			} catch {
-				container.textContent = content;
-			}
-		}).catch(() => {
-			if (!cancelled && container) {
-				container.textContent = content;
-			}
-		});
-
-		return () => {
-			cancelled = true;
-		};
-	}, [content]);
-
-	return <div ref={containerRef} className='positron-notebook-latex-output' />;
+	return (
+		<div ref={containerRef} className='positron-notebook-latex-output'>
+			<KatexMath displayMode latex={latex} />
+		</div>
+	);
 }
